+++ Basil/src/Scene/Scene.cpp
@@ -49,8 +49,8 @@ namespace Basil
 		return registry;
 	}
 
-	// On update function
-	void Scene::onUpdate(Timestep timeStep)
+	// On update function for runtime
+	void Scene::onUpdateRuntime(Timestep timeStep)
 	{
 		// Update scripts
 		{
@@ -126,6 +126,19 @@ namespace Basil
 		}
 
 		return {};
+	}
+
+	// On update function for the editor
+	void Scene::onUpdateEditor(Timestep timeStep, EditorCamera& camera)
+	{
+		Renderer2D::beginScene(camera);
+		auto group = registry.group<TransformComponent>(entt::get<SpriteRendererComponent>);
+		for (auto entity : group)
+		{
+			auto [transform, sprite] = group.get<TransformComponent, SpriteRendererComponent>(entity);
+			Renderer2D::drawQuad(transform.getTransform(), sprite.color);
+		}
+		Renderer2D::endScene();
 	}
 
 	// On component added (only called if component you are trying to add does not exist)